[Refactor #285] visual영역 분리
