+++ app/(withGlobalLayout)/main/page.tsx
@@ -1,8 +1,6 @@
 "use client";
 import PeopleIcon from "@/assets/images/icon_person.svg";
 import SearchIcon from "@/assets/images/icon_search.svg";
-import MafiaGameTitle from "@/assets/images/mafia_game_title.svg";
-import SongGameTitle from "@/assets/images/song_game_title.svg";
 import MafiaItem from "@/assets/images/mafia_item.png";
 import VisitEmptyImage from "@/assets/images/visit_empty.svg";
 import useConnectStore from "@/store/connect-store";
@@ -19,14 +17,10 @@ import { useEffect, useRef, useState } from "react";
 import { toast } from "react-toastify";
 import MainCreateRoom from "@/components/modal/CreateRoomModal";
 import { useCreateStore } from "../../../store/toggle-store";
-import { Swiper, SwiperSlide } from "swiper/react";
-import { EffectFade, Parallax, Autoplay, Pagination, Navigation } from "swiper/modules";
-import SwiperType from "swiper";
-import "swiper/css/bundle";
-import "swiper/css/pagination";
-import "swiper/css/navigation";
-import "@/style/mainpage/swiper.css";
 import useSocketOn from "@/hooks/useSocketOn";
+import MainVisual from "@/components/main/MainVisual";
+import RoomSearch from "@/utils/RoomSearch";
+import RoomListItem from "@/components/main/RoomListItem";
 
 const Mainpage = () => {
   const { isCreate, setIsCreate } = useCreateStore();
@@ -36,7 +30,6 @@ const Mainpage = () => {
   const isGoInClick = useRef(false);
   const roomId = useRef("");
   const router = useRouter();
-  const swiperRef = useRef<SwiperType | null>(null);
 
   useEffect(() => {
     socket.connect();
@@ -196,103 +189,17 @@ const Mainpage = () => {
     }
   };
 
-  const mouseEnterHandler = () => {
-    if (swiperRef.current) {
-      swiperRef.current.autoplay.stop();
-    }
-  };
-
-  const mouseLeaveHandler = () => {
-    if (swiperRef.current) {
-      swiperRef.current.autoplay.start();
-    }
-  };
-
-  //NOTE - 노래맞추기 버튼
-  const songHandler = () => {
-    toast("서비스 준비 중 입니다.");
-  };
-
   return (
     <main className={S.main}>
       <section className={S.visualSection}>
-        <Swiper
-          className={S.gameList}
-          pagination={{
-            type: "fraction"
-          }}
-          rewind={true}
-          effect={"fade"}
-          autoplay={{
-            delay: 5000,
-            disableOnInteraction: false
-          }}
-          navigation={true}
-          speed={800}
-          parallax={true}
-          modules={[EffectFade, Parallax, Autoplay, Pagination, Navigation]}
-          onSwiper={(swiper) => {
-            swiperRef.current = swiper;
-          }}
-        >
-          <SwiperSlide className={S.gameImage}>
-            <div
-              className={S.gameTitle}
-              data-swiper-parallax="-100%"
-              onMouseEnter={mouseEnterHandler}
-              onMouseLeave={mouseLeaveHandler}
-            >
-              <h2>
-                <Image src={MafiaGameTitle} alt="mafia game title" priority />
-              </h2>
-              <div className={S.gameButton}>
-                <button onClick={gameStartHandler}>Game Start</button>
-                <Link href="/mafiainfo" className={S.gameInfo}>
-                  More Info
-                </Link>
-              </div>
-            </div>
-          </SwiperSlide>
-          <SwiperSlide className={S.gameImage}>
-            <div
-              className={S.gameTitle}
-              data-swiper-parallax="200%"
-              onMouseEnter={mouseEnterHandler}
-              onMouseLeave={mouseLeaveHandler}
-            >
-              <h2>
-                <Image src={SongGameTitle} alt="song game title" priority />
-              </h2>
-              <div className={S.gameButton}>
-                <button onClick={songHandler}>Game Start</button>
-                <button onClick={songHandler} className={S.gameInfo}>
-                  More Info
-                </button>
-              </div>
-            </div>
-          </SwiperSlide>
-        </Swiper>
+        <MainVisual gameStartHandler={gameStartHandler} />
       </section>
       <div className={S.roomSectionWrap}>
         <section className={S.roomSection}>
           <div className={S.MainGnb}>
             <p>현재 활성화 되어있는 방</p>
             <div className={S.roomSearchAndButton}>
-              <form onSubmit={searchHandler}>
-                <div className={S.roomSearch}>
-                  <label htmlFor="RoomSearch">방 검색하기</label>
-                  <input
-                    type="text"
-                    id="RoomSearch"
-                    value={search}
-                    onChange={(e) => setSearch(e.target.value)}
-                    placeholder="방 이름을 입력해 주세요."
-                  />
-                  <button>
-                    <Image src={SearchIcon} alt="search Icon" />
-                  </button>
-                </div>
-              </form>
+              <RoomSearch />
               <div className={S.gameGoButton}>
                 <button disabled={isGoInClick.current} onClick={fastJoinRoomHandler}>
                   빠른입장
@@ -309,24 +216,7 @@ const Mainpage = () => {
           {rooms.length > 0 ? (
             <ul className={S.roomList}>
               {rooms.map((item) => (
-                <li key={item.room_id}>
-                  <Image src={MafiaItem} alt="room image" />
-                  <div className={S.roomTitle}>
-                    <h3>{item.title}</h3>
-                    <div className={S.gameName}>
-                      <p className={S.mafiaHashtag}>#&nbsp;{item.game_category}</p>
-                      <p className={S.currentPeople}>
-                        <Image src={PeopleIcon} alt="people icon" />
-                        <span>
-                          {item.current_user_count}/{item.total_user_count}
-                        </span>
-                      </p>
-                    </div>
-                  </div>
-                  <button disabled={isGoInClick.current} onClick={() => joinRoomHandler(item)} className={S.gotoButton}>
-                    입장하기
-                  </button>
-                </li>
+                <RoomListItem />
               ))}
             </ul>
           ) : (
